Struct creative_coding_project::app::audio::model::params::AudioParams

pub struct AudioParams {

    pub mask_resolution: Arc<Atomic<SpectralFilterSize>>,
    pub mask_is_post_fx: Arc<AtomicBool>,
    pub mask_mix: Arc<AtomicF64>,
    pub reso_bank_scale: Arc<Atomic<Scale>>,
    pub reso_bank_root_note: Arc<AtomicU8>,
    pub reso_bank_spread: Arc<SmootherAtomic<f64>>,
    pub reso_bank_shift: Arc<SmootherAtomic<f64>>,
    pub reso_bank_inharm: Arc<SmootherAtomic<f64>>,
    pub reso_bank_pan: Arc<SmootherAtomic<f64>>,
    pub reso_bank_quantize: Arc<AtomicBool>,
    pub reso_bank_resonator_count: Arc<AtomicU32>,
    pub reso_bank_mix: Arc<SmootherAtomic<f64>>,
    pub exciter_osc: Arc<Atomic<ExciterOscillator>>,
    pub low_filter_cutoff: Arc<SmootherAtomic<f64>>,
    pub low_filter_q: Arc<SmootherAtomic<f64>>,
    pub low_filter_gain_db: Arc<SmootherAtomic<f64>>,
    pub low_filter_is_shelf: Arc<AtomicBool>,
    pub high_filter_cutoff: Arc<SmootherAtomic<f64>>,
    pub high_filter_q: Arc<SmootherAtomic<f64>>,
    pub high_filter_gain_db: Arc<SmootherAtomic<f64>>,
    pub high_filter_is_shelf: Arc<AtomicBool>,
    pub delay_time_ms: Arc<AtomicF64>,
    pub delay_feedback: Arc<SmootherAtomic<f64>>,
    pub delay_mix: Arc<SmootherAtomic<f64>>,
    pub use_ping_pong: Arc<AtomicBool>,
    pub dist_amount: Arc<SmootherAtomic<f64>>,
    pub dist_type: Arc<Atomic<DistortionType>>,
    pub comp_thresh: Arc<SmootherAtomic<f64>>,
    pub comp_ratio: Arc<SmootherAtomic<f64>>,
    pub comp_attack_ms: Arc<SmootherAtomic<f64>>,
    pub comp_release_ms: Arc<SmootherAtomic<f64>>,
    pub pre_fx_gain: Arc<SmootherAtomic<f64>>,
    pub master_gain: Arc<SmootherAtomic<f64>>,

}

All audio-related parameters linked to the UI.

Fields
mask_resolution: Arc<Atomic<SpectralFilterSize>>

The block size of the spectral filter.

mask_is_post_fx: Arc<AtomicBool>

Whether the spectral filter is post-FX or not.

mask_mix: Arc<AtomicF64>reso_bank_scale: Arc<Atomic<Scale>>

The musical scale of the resonator bank.

reso_bank_root_note: Arc<AtomicU8>

The root note of the resonator bank.

reso_bank_spread: Arc<SmootherAtomic<f64>>

The frequency spread (range) of each resonator.

reso_bank_shift: Arc<SmootherAtomic<f64>>

The frequency shift of each resonator.

reso_bank_inharm: Arc<SmootherAtomic<f64>>

How much each resonator’s pitch skews towards its original pitch.

reso_bank_pan: Arc<SmootherAtomic<f64>>

How much panning may be applied to each resonator.

reso_bank_quantize: Arc<AtomicBool>

Whether the resonators should quantise their pitch to a scale.

reso_bank_resonator_count: Arc<AtomicU32>

The number of active resonators in the bank.

reso_bank_mix: Arc<SmootherAtomic<f64>>

The dry/wet mix of the resonator bank.

exciter_osc: Arc<Atomic<ExciterOscillator>>

The exciter oscillator.

low_filter_cutoff: Arc<SmootherAtomic<f64>>

The cutoff of the filter in Hz.

low_filter_q: Arc<SmootherAtomic<f64>>

The Q value of the cut filter.

low_filter_gain_db: Arc<SmootherAtomic<f64>>

The gain value of the shelf filter

low_filter_is_shelf: Arc<AtomicBool>

Whether the low filter is a shelf filter or not.

high_filter_cutoff: Arc<SmootherAtomic<f64>>

The cutoff of the high-pass filter in Hz.

high_filter_q: Arc<SmootherAtomic<f64>>

The Q value of the high-pass filter.

high_filter_gain_db: Arc<SmootherAtomic<f64>>

The gain value of the shelf filter.

high_filter_is_shelf: Arc<AtomicBool>

Whether the high filter is a shelf filter or not.

delay_time_ms: Arc<AtomicF64>

The time between delay taps in milliseconds.

delay_feedback: Arc<SmootherAtomic<f64>>

The delay feedback.

delay_mix: Arc<SmootherAtomic<f64>>

The dry/wet mix for the delay.

use_ping_pong: Arc<AtomicBool>

Whether to use ping-pong delay or not.

dist_amount: Arc<SmootherAtomic<f64>>

The “amount” of distortion, i.e. drive.

dist_type: Arc<Atomic<DistortionType>>

The distortion algorithm.

comp_thresh: Arc<SmootherAtomic<f64>>

Compressor threshold.

comp_ratio: Arc<SmootherAtomic<f64>>

Compressor ratio.

comp_attack_ms: Arc<SmootherAtomic<f64>>

Compressor attack time in milliseconds.

comp_release_ms: Arc<SmootherAtomic<f64>>

Compressor release time in milliseconds.

pre_fx_gain: Arc<SmootherAtomic<f64>>master_gain: Arc<SmootherAtomic<f64>>

The device’s master gain level.