Struct creative_coding_project::app::params::ui::UIParams

pub struct UIParams {

    pub mask_algorithm: Arc<Atomic<GenerativeAlgo>>,
    pub mask_scan_line_speed: Arc<AtomicF64>,
    pub mask_is_post_fx: Arc<AtomicBool>,
    pub mask_mix: Arc<AtomicF64>,
    pub mask_resolution: Arc<Atomic<SpectralFilterSize>>,
    pub contour_count: Arc<AtomicU32>,
    pub contour_thickness: Arc<AtomicF64>,
    pub contour_speed: Arc<AtomicF64>,
    pub smoothlife_resolution: Arc<Atomic<SmoothLifeSize>>,
    pub smoothlife_speed: Arc<AtomicF64>,
    pub smoothlife_preset: Arc<Atomic<SmoothLifePreset>>,
    pub spectrogram_resolution: Arc<Atomic<SpectrogramSize>>,
    pub spectrogram_timing: Arc<AtomicF64>,
    pub spectrogram_view: Arc<Atomic<SpectrogramView>>,
    pub voronoi_cell_count: Arc<AtomicU32>,
    pub voronoi_cell_speed: Arc<AtomicF64>,
    pub voronoi_border_weight: Arc<AtomicF64>,
    pub reso_bank_scale: Arc<Atomic<Scale>>,
    pub reso_bank_root_note: Arc<AtomicU8>,
    pub reso_bank_spread: Arc<SmootherAtomic<f64>>,
    pub reso_bank_shift: Arc<SmootherAtomic<f64>>,
    pub reso_bank_inharm: Arc<SmootherAtomic<f64>>,
    pub reso_bank_pan: Arc<SmootherAtomic<f64>>,
    pub reso_bank_quantize: Arc<AtomicBool>,
    pub reso_bank_resonator_count: Arc<AtomicU32>,
    pub reso_bank_cell_count: Arc<AtomicU32>,
    pub reso_bank_cell_jitter: Arc<AtomicF64>,
    pub reso_bank_field_friction: Arc<AtomicF64>,
    pub reso_bank_mix: Arc<SmootherAtomic<f64>>,
    pub exciter_osc: Arc<Atomic<ExciterOscillator>>,
    pub low_filter_cutoff: Arc<SmootherAtomic<f64>>,
    pub low_filter_q: Arc<SmootherAtomic<f64>>,
    pub low_filter_gain_db: Arc<SmootherAtomic<f64>>,
    pub low_filter_is_shelf: Arc<AtomicBool>,
    pub high_filter_cutoff: Arc<SmootherAtomic<f64>>,
    pub high_filter_q: Arc<SmootherAtomic<f64>>,
    pub high_filter_gain_db: Arc<SmootherAtomic<f64>>,
    pub high_filter_is_shelf: Arc<AtomicBool>,
    pub delay_time_ms: Arc<AtomicF64>,
    pub delay_feedback: Arc<SmootherAtomic<f64>>,
    pub delay_mix: Arc<SmootherAtomic<f64>>,
    pub use_ping_pong: Arc<AtomicBool>,
    pub dist_amount: Arc<SmootherAtomic<f64>>,
    pub dist_type: Arc<Atomic<DistortionType>>,
    pub comp_thresh: Arc<SmootherAtomic<f64>>,
    pub comp_ratio: Arc<SmootherAtomic<f64>>,
    pub comp_attack_ms: Arc<SmootherAtomic<f64>>,
    pub comp_release_ms: Arc<SmootherAtomic<f64>>,
    pub pre_fx_gain: Arc<SmootherAtomic<f64>>,
    pub master_gain: Arc<SmootherAtomic<f64>>,

}

All parameters controlled by the user interface.

Fields
mask_algorithm: Arc<Atomic<GenerativeAlgo>>

The algorithm to use for the spectral filter.

mask_scan_line_speed: Arc<AtomicF64>

The speed of the spectral filter scan line.

mask_is_post_fx: Arc<AtomicBool>

Whether the spectral filter is pre- or post-FX.

mask_mix: Arc<AtomicF64>mask_resolution: Arc<Atomic<SpectralFilterSize>>

The block size of the spectral filter.

contour_count: Arc<AtomicU32>

The number of contours to draw.

contour_thickness: Arc<AtomicF64>

The thickness of each contour line.

contour_speed: Arc<AtomicF64>

The speed of the contour animation.

smoothlife_resolution: Arc<Atomic<SmoothLifeSize>>

The resolution of the smooth life simulation.

smoothlife_speed: Arc<AtomicF64>

The speed of the smoothlife simulation.

smoothlife_preset: Arc<Atomic<SmoothLifePreset>>

The state preset of the smoothlife simulation.

spectrogram_resolution: Arc<Atomic<SpectrogramSize>>

The resolution of both spectrograms.

spectrogram_timing: Arc<AtomicF64>

The timing of the spectrograms.

spectrogram_view: Arc<Atomic<SpectrogramView>>

Which spectrograms are drawn.

voronoi_cell_count: Arc<AtomicU32>

The number of Voronoi cells.

voronoi_cell_speed: Arc<AtomicF64>

The speed of the Voronoi flow field.

voronoi_border_weight: Arc<AtomicF64>

The weight of the Voronoi borders and isolines.

reso_bank_scale: Arc<Atomic<Scale>>

The musical scale of the resonator bank.

reso_bank_root_note: Arc<AtomicU8>

The root note of the resonator bank.

reso_bank_spread: Arc<SmootherAtomic<f64>>

The frequency spread (range) of each resonator.

reso_bank_shift: Arc<SmootherAtomic<f64>>

The frequency shift of each resonator.

reso_bank_inharm: Arc<SmootherAtomic<f64>>

How much each resonator’s pitch skews towards its original pitch.

reso_bank_pan: Arc<SmootherAtomic<f64>>

How much panning may be applied to each resonator.

reso_bank_quantize: Arc<AtomicBool>

Whether the resonators should quantise their pitch to a scale.

reso_bank_resonator_count: Arc<AtomicU32>

The number of active resonators in the bank.

reso_bank_cell_count: Arc<AtomicU32>reso_bank_cell_jitter: Arc<AtomicF64>reso_bank_field_friction: Arc<AtomicF64>

The friction applied to each point in the vector field.

reso_bank_mix: Arc<SmootherAtomic<f64>>

The dry/wet mix of the resonator bank.

exciter_osc: Arc<Atomic<ExciterOscillator>>

The exciter oscillator.

low_filter_cutoff: Arc<SmootherAtomic<f64>>

The cutoff of the filter in Hz.

low_filter_q: Arc<SmootherAtomic<f64>>

The Q value of the cut filter.

low_filter_gain_db: Arc<SmootherAtomic<f64>>

The gain value of the shelf filter

low_filter_is_shelf: Arc<AtomicBool>

Whether the low filter is a shelf filter or not.

high_filter_cutoff: Arc<SmootherAtomic<f64>>

The cutoff of the high-pass filter in Hz.

high_filter_q: Arc<SmootherAtomic<f64>>

The Q value of the high-pass filter.

high_filter_gain_db: Arc<SmootherAtomic<f64>>

The gain value of the shelf filter.

high_filter_is_shelf: Arc<AtomicBool>

Whether the high filter is a shelf filter or not.

delay_time_ms: Arc<AtomicF64>

The time between delay taps in milliseconds.

delay_feedback: Arc<SmootherAtomic<f64>>

The delay feedback.

delay_mix: Arc<SmootherAtomic<f64>>

The dry/wet mix for the delay.

use_ping_pong: Arc<AtomicBool>

Whether to use ping-pong delay or not.

dist_amount: Arc<SmootherAtomic<f64>>

The “amount” of distortion.

dist_type: Arc<Atomic<DistortionType>>

The distortion algorithm.

comp_thresh: Arc<SmootherAtomic<f64>>

Compression threshold in decibels.

comp_ratio: Arc<SmootherAtomic<f64>>

Compression ratio.

comp_attack_ms: Arc<SmootherAtomic<f64>>

Compression attack time in milliseconds.

comp_release_ms: Arc<SmootherAtomic<f64>>

Compression release time in milliseconds.

pre_fx_gain: Arc<SmootherAtomic<f64>>master_gain: Arc<SmootherAtomic<f64>>

Master gain level in decibels.