Struct creative_coding_project::app::Model

pub struct Model {

    pub ui_params: UIParams,
    pub audio_stream: Stream<AudioModel>,
    pub audio_senders: Arc<AudioMessageSenders>,
    pub spectral_mask: Input<SpectralMask>,
    pub voice_event_sender: Sender<VoiceEvent>,
    pub audio_callback_timer: Arc<Mutex<Instant>>,
    pub sample_rate_ref: Arc<AtomicF64>,
    pub octave: Octave,
    pub note_handler: NoteHandlerRef,
    pub pressed_keys: HashMap<Key, bool>,
    pub pre_spectrum_analyzer: SpectrumAnalyzer,
    pub post_spectrum_analyzer: SpectrumAnalyzer,
    pub ui_components: UIComponents,
    pub bank_rect: Rect,
    pub reso_bank_reset_receiver: Receiver<()>,
    pub reso_bank_reset_sender_key: Sender<()>,
    pub reso_bank_push_receiver: Receiver<()>,
    pub reso_bank_push_sender_key: Sender<()>,
    pub reso_bank_data: Input<ResoBankData>,
    pub mask_rect: Rect,
    pub mouse_clicked_outside_of_mask: bool,
    pub mask_clicked: bool,
    pub spectrum_rect: Rect,
    pub contours: Arc<RwLock<ContoursGPU>>,
    pub smooth_life: Arc<RwLock<SmoothLifeGPU>>,
    pub voronoi_mask: Arc<RwLock<VoronoiGPU>>,
    pub voronoi_vectors: Arc<RwLock<VectorField>>,
    pub vectors_reso_bank: VectorField,
    pub voronoi_reso_bank: VoronoiGPU,
    pub mask_scan_line_pos: f64,
    pub mask_scan_line_increment: f64,
    pub eq_display: EQDisplay,
    pub log_lines: Vec<[Vec2; 2]>,
    pub input_data: InputData,
    /* private fields */

}

The app’s model, i.e. its state.

Fields
ui_params: UIParams

UI parameters.

audio_stream: Stream<AudioModel>

The CPAL audio stream.

audio_senders: Arc<AudioMessageSenders>

Channels to send messages directly to the audio thread.

spectral_mask: Input<SpectralMask>

Input to the spectral mask channel, sent to the audio thread.

voice_event_sender: Sender<VoiceEvent>

Channel to send voice events (such as killing all voices).

audio_callback_timer: Arc<Mutex<Instant>>

A thread-safe reference to the timer which tracks when the audio callback was last called.

sample_rate_ref: Arc<AtomicF64>

A reference to the sample rate value.

octave: Octave

Current octave for note input (via typing keyboard).

note_handler: NoteHandlerRef

A thread-safe reference to the note handler.

pressed_keys: HashMap<Key, bool>

A HashMap of the currently-pressed keys.

pre_spectrum_analyzer: SpectrumAnalyzer

The pre-FX spectrogram.

post_spectrum_analyzer: SpectrumAnalyzer

The post-FX spectrogram.

ui_components: UIComponents

All GUI control components (sliders, menus, buttons, labels…)

bank_rect: Rect

The resonator bank bounding rect.

reso_bank_reset_receiver: Receiver<()>

A receiver for the resonator bank “reset” button.

reso_bank_reset_sender_key: Sender<()>

A sender to allow a keybind to reset the resonator bank.

reso_bank_push_receiver: Receiver<()>

A receiver for the resonator bank “push” button.

reso_bank_push_sender_key: Sender<()>

A sender to allow a keybind to push the resonator bank.

reso_bank_data: Input<ResoBankData>

Input to the resonator bank state channel, sent to the audio thread.

mask_rect: Rect

The spectral filter bounding rect.

mouse_clicked_outside_of_mask: bool

Whether the mouse was clicked outside of the spectral filter — used to track input events applied to the mask scan line.

mask_clicked: bool

Whether the spectral filter is clicked or not.

spectrum_rect: Rect

The spectrogram/EQ bounding rect.

contours: Arc<RwLock<ContoursGPU>>

A Perlin noise contour generator.

smooth_life: Arc<RwLock<SmoothLifeGPU>>

A SmoothLife simulation.

voronoi_mask: Arc<RwLock<VoronoiGPU>>

A Voronoi noise generator used for the spectral mask.

voronoi_vectors: Arc<RwLock<VectorField>>

A vector field used to manage points for the Voronoi mask.

vectors_reso_bank: VectorField

A simple vector field for the resonator bank points.

voronoi_reso_bank: VoronoiGPU

The voronoi generator for the resonator bank vector field.

mask_scan_line_pos: f64

The line which shows which column is being used as a spectral mask.

mask_scan_line_increment: f64

The amount to increment the position of the mask scan line each frame.

eq_display: EQDisplay

The EQ display — filter nodes and the frequency response line.

log_lines: Vec<[Vec2; 2]>

Logarithmic frequency lines drawn in the background of the spectrogram.

input_data: InputData

User input data.

Implementations

impl Model

pub fn build(app: &App) -> Self

Builds the app’s Model.

Panics

Panics if a new window cannot be initialized.

pub fn current_sample_idx(&self) -> u32

Returns the (approximate) sample index for the current moment in time.

This is not a particularly precise method of tracking time events, but should be more than adequate for things like note events.

If a lock on the callback timer is not obtained, then 0 is returned. This doesn’t create too much of an issue as note events are still handled quite quickly in the audio thread.

pub fn increment_mask_scan_line(&mut self)

Increments the position of the mask scan line.

pub fn mask_rect(&self) -> Rect

Panics

This will panic if the SmoothLife generator cannot be locked.

pub fn draw_mask_scan_line(&self, draw: &Draw)

Draws the spectral mask scan line.

pub fn update_input_data(&mut self, app: &App)

Updates the model’s input data.

pub fn update_reso_bank_vector_field(&mut self, app: &App)

Updates the resonator bank vector field.

pub fn update_voronoi_vectors(&mut self, app: &App)

Updates the vector field used for the Voronoi noise spectral mask.

pub fn update_mask_scan_line_from_mouse(&mut self)

Updates the mask scan line based on user mouse input.

pub fn update_eq(&mut self, app: &App)

Updates the EQ GUI.

pub fn draw_log_lines(&self, draw: &Draw)

Draws the spectrogram log lines.

pub fn reso_bank_needs_redraw(&self) -> bool

Determins whether the resonator bank needs to be redrawn.

TODO: for some reason the resonator bank vector field and voronoi noise do not update immediately, and it seems like the voronoi generator updates a frame or two (or three?) later. For now, this method is just used to determine whether to update the voronoi generator each frame, and not when to draw it.

pub fn redraw_under_menus(&self, draw: &Draw, is_first_frame: bool)

Redraws components under expanded when required.

pub fn redraw_filter_sliders(&self, draw: &Draw)

Redraws the filter text sliders when filter types are changed.