Struct creative_coding_project::dsp::filtering::resonator::resonator_bank::ResonatorBank

pub struct ResonatorBank { /* private fields */ }

Implementations

impl ResonatorBank

pub const NOTE_MIN: f64 = 30f64

pub const NOTE_MIDDLE: f64 = 79f64

pub const NOTE_MAX: f64 = 128f64

pub fn new(sample_rate: f64, max_num_resonators: usize) -> Self

pub fn set_params(&mut self, params: ResonatorBankParams)

pub fn set_num_resonators(&mut self, num_resonators: usize)

Sets the number of active resonators in the bank.

Panics

Panics if num_resonators > self.max_num_resonators().

pub fn max_num_resonators(&self) -> usize

Returns the maximum number of resonators available in the bank.

pub fn randomize(&mut self)

Randomizes the raw pitches of the resonator bank.

pub fn quantize_to_scale(&mut self, quantize_to_scale: bool)

pub fn set_panning_scale(&mut self, scale: f64)

Scales the panning applied to each resonator.

Clamped to [0.0 - 1.0].

pub fn set_freq_spread(&mut self, spread: f64)

pub fn set_freq_shift(&mut self, shift: f64)

pub fn set_root_note(&mut self, root_note_midi: f64)

Sets the root note of the bank’s scale.

Only active if quantise_to_scale is true.

pub fn set_scale(&mut self, scale: Scale)

Sets the internal musical scale of the bank.

Only active if quantise_to_scale is true.

pub fn set_inharm(&mut self, inharm: f64)

Sets how much each resonator pitch skews towards its original pitch.

Only active if quantise_to_scale is true.

pub fn original_pitches_mut(&mut self) -> &mut [f64]

Returns a mutable reference to the raw resonator pitches.

pub fn pan_values_mut(&mut self) -> &mut [Smoother<f64>]

Returns a mutable reference to the raw pan values.

Each value is controlled by a smoother, so use the set_target_value() method to change the value.

pub fn set_state_from_data(&mut self, data: &ResoBankData)

pub fn inner(&self) -> &[AudioUtility<StereoWrapper<TwoPoleResonator>>]

Returns a reference to the internal resonators.

pub fn inner_mut( &mut self ) -> &mut [AudioUtility<StereoWrapper<TwoPoleResonator>>]

Returns a mutable reference to the internal resonators.