Struct creative_coding_project::dsp::spectral::stft::helper::StftHelper

pub struct StftHelper { /* private fields */ }

A convenience struct for a short-time Fourier transform algorithm.

Implementations

impl StftHelper

pub fn new( num_channels: usize, max_block_size: usize, max_padding: usize ) -> Self

Returns a new StftHelper.

Panics

Panics if num_channels or max_block_size is 0.

pub fn set_block_size(&mut self, block_size: usize)

Sets the block size. This clears the internal buffers, meaning the next block will be silent.

Panics

Panics if block_size > max_block_size.

pub fn clear(&mut self)

Clears the internal buffers.

pub fn set_padding(&mut self, padding: usize)

Set the amount of padding. This clears the internal buffers, meaning the next block will be silent.

Panics

Panics if padding > max_padding.

pub fn num_channels(&self) -> usize

Number of channels the StftHelper was set up with.

pub fn max_block_size(&self) -> usize

Maximum block size the StftHelper was set up with.

pub fn max_padding(&self) -> usize

Maximum amount of padding the StftHelper was set up with.

pub fn latency_samples(&self) -> u32

Amount of latency produced by the STFT process in samples.

pub fn process_overlap_add<M, F>( &mut self, main_buffer: &mut M, overlap_factor: usize, callback: F )where M: StftInputMut, F: FnMut(usize, &mut [f64]),

Processes the audio from main_buffer in short, overlapping blocks, then sums the blocks back to the main buffer one block later. This means that this function introduces one block of latency. Use latency_samples() to obtain the amount of latency in samples.

If a padding value was provided to the StftHelper, that many zeroes will be appended to each block. Padding values are added to the next block before the callback is called.

Any window functions must be applied within the callback, as there are many to choose from.

The callback function expects its first argument to be the channel index (as usize), and its second to be the real-valued buffer (as &mut [f64]). The real-valued buffer is a slice of block_size read-valued samples, which can be passed directly to most FFT algorithms.

This function reuses the same buffer for all calls to callback, which does mean that you can only access one channel of windowed data at a time.

Panics

Panics if the buffer has a different number of channels to the StftHelper, or if overlap_factor == 0.

pub fn process_forward_only<B, F>( &mut self, main_buffer: &B, overlap_factor: usize, callback: F )where B: StftInput, F: FnMut(usize, &mut [f64]),

Only processes the forward FFT, so the buffer is only ever read, not mutated.

The callback function expects its first argument to be the channel index (as usize), and its second to be the real-valued buffer (as &mut [f64]).

Panics

Panics if the buffer has a different number of channels to the StftHelper, or if overlap_factor == 0.