Trait creative_coding_project::dsp::spectral::stft::stft_trait::StftInput

pub trait StftInput {
    // Required methods
    fn num_samples(&self) -> usize;
    fn num_channels(&self) -> usize;
    unsafe fn get_sample_unchecked(
        &self,
        channel_idx: usize,
        sample_idx: usize
    ) -> f64;
}

A buffer which may be read by the StftHelper.

Required Methods

fn num_samples(&self) -> usize

Number of samples in the input buffer.

fn num_channels(&self) -> usize

Number of channels in the input buffer.

unsafe fn get_sample_unchecked( &self, channel_idx: usize, sample_idx: usize ) -> f64

Obtains a copy of a specific sample without any bounds checking.