Struct creative_coding_project::generative::contours::Contours

pub struct Contours { /* private fields */ }

Perlin noise contour line generator. Supports multi-threading.

Implementations

impl Contours

pub fn new(device: &Device, rect: Rect) -> Self

Creates a new Contours.

This object is responsible for handling its own texture and image buffer.

Uses 1 thread by default — see with_num_threads().

pub fn randomize(&mut self)

Randomizes the internal z value.

pub fn with_contour_range(self, range: RangeInclusive<f64>) -> Self

Adds the provided range to self.

This is the range in which pixels are drawn for each contour.

Has no effect if the range is outside 0 <= x <= 1.

Consuming method.

pub fn set_contour_range(&mut self, range: RangeInclusive<f64>)

Sets the range in which pixels are drawn for each contour.

Has no effect if the range is outside 0 <= x <= 1.

pub fn with_num_contours(self, num_contours: u32) -> Self

Sets how many contours to compute.

Consuming method.

pub fn set_num_contours(&mut self, num_contours: u32)

Sets how many contours to compute.

pub fn with_z_increment(self, z_increment: f64) -> Self

Sets how much to increment the z value per frame to transition through a third noise dimension.

Consuming method.

pub fn set_z_increment(&mut self, z_increment: f64)

Sets how much to increment the z value per frame to transition through a third noise dimension.

pub fn with_num_threads(self, num_threads: usize) -> Option<Self>

Sets the number of threads to use for computing the noise contours.

This method will allocate if the threads are spawned.

Consuming method. Returns Some if the threads were successfully spawned, and None otherwise.

# Panics

Panics if num_threads is not a power-of-two value (required to efficiently divide the image buffer up).

pub fn set_num_threads(&mut self, num_threads: usize) -> bool

Sets the number of threads to use for computing the noise contours.

Returns true if the threads were successfully spawned, and false otherwise.

This method will allocate if new threads are successfully spawned.

Panics

Panics if num_threads is not a power-of-two value (required to efficiently divide the image buffer up).

pub fn with_feathering(self, should_feather: bool) -> Self

Controls whether the contour lines should be feathered or not.

Consuming method.

pub fn set_feathering(&mut self, should_feather: bool)

Controls whether the contour lines should be feathered or not.

pub fn width_px(&self) -> usize

The width of the image buffer in pixels.

pub fn height_px(&self) -> usize

The height of the image buffer in pixels.