Struct creative_coding_project::gui::components::menu::Menu

pub struct Menu<E: MenuEnum> {
    pub needs_redraw: bool,
    /* private fields */
}

A menu UI component, which essentially exposes an enum as a menu.

Requires that the enum implements the MenuEnum trait.

Fields
needs_redraw: bool
Implementations

impl<E: MenuEnum> Menu<E>

pub fn new(rect: Rect) -> Self

Creates a new, default Menu.

pub fn with_label(self, label: &str) -> Self

Provides the Menu with a label.

pub fn with_label_layout(self, layout: Layout) -> Self

Sets the label’s text layout.

pub fn with_label_font_size(self, size: u32) -> Self

Sets the font size for the menu label.

pub fn with_item_text_layout(self, layout: Layout) -> Self

Sets each item’s text layout.

pub fn with_item_font_size(self, size: u32) -> Self

Sets the font size for the item labels.

pub fn with_font_size(self, size: u32) -> Self

Sets the font size for the whole Menu.

pub fn with_selected_item_text_layout(self, layout: Layout) -> Self

Provides a text layout for the selected item when the menu is open.

pub fn with_callback<F>(self, callback: F) -> Selfwhere F: Fn(E) + 'static,

Provides a callback to the menu, which is called whenever a new item is selected.

The callback’s argument is the current variant of the menu.

pub fn starts_open(self) -> Self

Creates the menu in its open state.

pub fn initial_variant(self, variant: E) -> Self

Sets the initial item variant for the menu.

pub fn set_callback<F: Fn(E) + 'static>(&mut self, callback: F)

Sets a callback, which is called whenever a new item is selected.

The callback’s argument is the current variant of the menu.

pub fn output(&self) -> E

Returns the current item of the menu.

pub fn label_layout(&self) -> &Layout

Returns a reference to the menu’s label text layout.

pub fn item_text_layout(&self) -> &Layout

Returns a reference to the menu’s item text layout.

pub fn selected_text_layout(&self) -> Option<&Layout>

Returns a reference to the menu’s selected item text layout.

pub fn label_layout_mut(&mut self) -> &mut Layout

Returns a mutable reference to the menu’s label text layout.

pub fn item_text_layout_mut(&mut self) -> &mut Layout

Returns a mutable reference to the menu’s item text layout.

pub fn selected_text_layout_mut(&mut self) -> Option<&mut Layout>

Returns a mutable reference to the menu’s selected item text layout.

pub fn is_open(&self) -> bool

Returns whether the menu is open or not.

pub fn redraw_label(&self, draw: &Draw)