Trait creative_coding_project::gui::components::menu_trait::MenuEnum

pub trait MenuEnum: Display + Clone + Copy + Default + Debug {
    // Required methods
    fn num_variants() -> usize;
    fn idx(&self) -> usize;
    fn from_idx(idx: usize) -> Option<Self>;
}

A trait for preparing an enum for use with the [Menu] UI component.

The enum must also implement:

Display, so its name can be printed in the menu

Clone and Copy, so the value can be returned from the menu

Default, so the menu knows which variant to use when created

Debug, so the whole menu can be debugged

Please see the documentation of all the required methods for more information.

Required Methods

fn num_variants() -> usize

The number of variants in the enum, used to layout the menu.

Note: you can technically set this to less than the actual number of variants to truncate the menu. Setting a value greater than the number of variants, however, will cause the menu to panic when it is created.

fn idx(&self) -> usize

The index for a given variant. Used to avoid having to iterate through all variants each frame. You should ensure that this matches the layout of from_idx().

fn from_idx(idx: usize) -> Option<Self>

The variant from a given index. This method should be used to describe the order of elements in the menu.

Object Safety

This trait is not object safe.