Struct creative_coding_project::gui::components::text_slider::TextSlider

pub struct TextSlider { /* private fields */ }

A simple slider with a text readout.

Holding shift whilst dragging the slider enables 10x fine value control; alt- or cmd-clicking the slider will reset it to its default value, if one is provided.

Scrolling over the value will also change it.

TODO: implement logarithmic scaling. TODO: implement coloring. TODO: implement the ability to manually type and parse a value. (yikes)

Implementations

impl TextSlider

pub fn new(value: f64, rect: Rect) -> Self

Created a new TextSlider with default options.

pub fn with_label(self, label: &str) -> Self

Provides the TextSlider with a label.

pub fn with_font_size(self, size: u32) -> Self

Sets the font size for both the label and value readout.

pub fn with_label_layout(self, layout: Layout) -> Self

Sets the text layout of the label.

pub fn with_value_layout(self, layout: Layout) -> Self

Sets the text layout of the value readout.

pub fn with_value_chars(self, num_chars: usize) -> Self

Sets the number of chars used to show the value read out. The default value is 5. This includes any decimal points, but ignores any prefix or suffix. Does not apply if the slider uses integer rounding.

Note that this value is not strictly adhered to, and the text formatting will always allow the whole value to be shown, and always includes 1 decimal value if the value would otherwise be truncated at the decimal point.

See with_max_value_num_chars() to have use a constant number of chars.

pub fn with_max_value_chars(self) -> Self

Sets the number of chars used to show the value readout based on the maximum value of the slider’s range, such that the number of chars in the readout will never change. Does not apply if the slider uses integer rounding.

pub fn with_output_range(self, range: RangeInclusive<f64>) -> Self

Sets the output range of the TextSlider. The default is 0.0..=1.0.

pub fn with_default_value(self, value: f64) -> Self

Provides a default value to the TextSlider.

Must be called after with_output_range(), if provided.

pub fn with_log_scaling(self) -> Self

Enables logarithmic output value scaling for the TextSlider.

pub fn with_prefix(self, prefix: &str) -> Self

Provides a text prefix for the value readout.

pub fn with_suffix(self, suffix: &str) -> Self

Provides a text suffix for the value readout.

pub fn with_positive_value_prefix(self) -> Self

Prefixes positive values with “+”.

pub fn with_callback<F>(self, cb: F) -> Selfwhere F: Fn(f64, f64) + 'static,

Provides a callback which is called whenever the TextSlider’s value is updated.

The first argument is the raw slider value, and the second is the output value.

pub fn with_formatting_callback<F: Fn(f64, f64) -> String + 'static>( self, cb: F ) -> Self

Provides a callback to format the value text. Overrides any text-formatting options passed to the TextSlider when provided.

The first argument is the raw slider value, and the second is the output value.

pub fn with_sensitivity(self, sensitivity: f64) -> Self

Sets the drag sensitivity of the TextSlider. The default value is 0.004.

pub fn with_integer_rounding(self) -> Self

Uses integer rounding for the output value.

pub fn set_label(&mut self, label: &str)

Provides the TextSlider with a label.

pub fn set_font_size(&mut self, size: u32)

Sets the font size both the label and value readout. If you want to control these font sizes separately, see the label_layout_mut() and value_layout_mut() methods.

pub fn set_label_layout(&mut self, layout: Layout)

Sets the text layout of the label.

pub fn set_value_layout(&mut self, layout: Layout)

Sets the text layout of the value readout.

pub fn set_output_range(&mut self, range: RangeInclusive<f64>)

Sets the output range of the TextSlider. The default is 0.0..=1.0.

pub fn set_default_value(&mut self, value: f64)

Sets the default value of the TextSlider.

pub fn log_scaling(&mut self, use_log_scaling: bool)

Sets whether to use logarithmic scaling for the output value.

pub fn set_prefix(&mut self, prefix: &str)

Sets the prefix for the value readout.

pub fn set_suffix(&mut self, suffix: &str)

Sets the suffix for the value readout.

pub fn set_positive_prefix(&mut self, prefix: bool)

Sets whether to prefix positive values with “+”.

pub fn set_value_num_chars(&mut self, num_chars: usize)

Sets the number of chars used to represent the value read out. The default value is 5.

This includes any decimal points, but ignores any prefix or suffix.

pub fn set_max_value_chars(&mut self)

Sets the number of chars used to show the value readout based on the maximum value of the slider’s range, such that the number of chars in the readout will never change. Does not apply if the slider uses integer rounding.

pub fn set_sensitivity(&mut self, sensitivity: f64)

Sets the drag sensitivity of the TextSlider. The default value is 0.001.

pub fn set_callback<F: Fn(f64, f64) + 'static>(&mut self, cb: F)

Provides a callback which is called whenever the TextSlider’s value is updated.

The first argument is the raw slider value, and the second is the output value.

pub fn detach_callback(&mut self)

Detaches any attached callback from the TextSlider.

See set_callback() for more information.

pub fn set_formatting_callback<F: Fn(f64, f64) -> String + 'static>( &mut self, cb: F )

Provides a callback to format the value text. Overrides any text-formatting options passed to the TextSlider when provided.

pub fn detach_formatting_callback(&mut self)

Detaches any attached formatting callback from the TextSlider.

See set_formatting_callback() for more information.

pub fn set_integer_rounding(&mut self, use_integer_rounding: bool)

Uses integer rounding for the output value.

pub fn label_layout(&self) -> &Layout

Returns a reference to the current label text layout.

pub fn label_layout_mut(&mut self) -> &mut Layout

Returns a mutable reference to the current label text layout.

pub fn value_layout(&self) -> &Layout

Returns a reference to the current value text layout.

pub fn value_layout_mut(&mut self) -> &mut Layout

Returns a mutable reference to the current value text layout.

pub fn is_active(&self) -> bool

Returns whether the TextSlider is current active or not — this may act as a way of checking whether the TextSlider should be updated or not.

pub fn value(&self) -> f64

Returns the output value of the TextSlider.

pub fn value_raw(&self) -> f64

Returns the raw, normalised value of the TextSlider.

pub fn set_value(&mut self, value: f64)

Sets the output value of the TextSlider.

pub fn set_value_raw(&mut self, value: f64)

Sets the raw value of the TextSlider.

pub fn within_bounds(&self, pos: Vec2) -> bool

Returns whether pos is contained within the bounding rect of the TextSlider.