Trait creative_coding_project::gui::draw_traits::DrawMask

pub trait DrawMask: UIDraw {
    // Provided methods
    fn column_to_mask(&self, mask: &mut SpectralMask, len: usize, x: f64) { ... }
    fn row_to_mask(&self, mask: &mut SpectralMask, len: usize, y: f64) { ... }
}

Trait for UI components which act as spectral masks.

Provided Methods

fn column_to_mask(&self, mask: &mut SpectralMask, len: usize, x: f64)

A method to map columnar data from self to a SpectralMask.

fn row_to_mask(&self, mask: &mut SpectralMask, len: usize, y: f64)

A method to map tabular data from self to a SpectralMask.