Trait creative_coding_project::gui::draw_traits::UIDraw

pub trait UIDraw {
    // Required methods
    fn update(&mut self, app: &App, input_data: &InputData);
    fn draw(&self, app: &App, draw: &Draw, frame: &Frame<'_>);
    fn rect(&self) -> &Rect;

    // Provided methods
    fn should_update(&self, mouse_data: &InputData) -> bool { ... }
    fn needs_redraw(&self) -> bool { ... }
    fn force_redraw(&mut self, app: &App, draw: &Draw, frame: &Frame<'_>) { ... }
    fn draw_bounding_rect(&self, draw: &Draw) { ... }
}

Trait for UI components which can be drawn. See each method for more information.

Required Methods

fn update(&mut self, app: &App, input_data: &InputData)

The component’s update method, to be used in the app’s update() callback.

This method should update any internal state based on the time and input events.

fn draw(&self, app: &App, draw: &Draw, frame: &Frame<'_>)

The component’s draw loop, to be called in the app’s view() callback.

This method should only draw the component, not update any state.

fn rect(&self) -> &Rect

Returns a reference to the component’s bounding rect.

Provided Methods

fn should_update(&self, mouse_data: &InputData) -> bool

An optional method for querying whether the component should be updated or not.

To defer control to the component, it is recommended that this method is used as a return guard in the component’s update() method.

fn needs_redraw(&self) -> bool

An optional method for querying whether the component needs to be redrawn or not. Returns true as default.

To defer control to the component, it is recommended that this method is used as a return guard in the component’s draw() method.

fn force_redraw(&mut self, app: &App, draw: &Draw, frame: &Frame<'_>)

An optional method to force the component to redraw.

fn draw_bounding_rect(&self, draw: &Draw)

A method for drawing the component’s bounding rect. Useful for debugging.

By default, this will draw an outline of the rect returned by the rect() method.