Module creative_coding_project::prelude

Re-exports

pub use crate::app::audio::NoteEvent;

pub use crate::app::Scale;

pub use crate::gui::DrawMask;

pub use crate::gui::InputData;

pub use crate::gui::UIDraw;

pub use crate::simd::SimdBuffer;

pub use crate::simd::SimdType;

pub use crate::settings::*;

pub use crate::util::*;

Structs

AtomicF64

A floating point type which can be safely shared between threads.

DVec2

A 2-dimensional vector.

Vec2

A 2-dimensional vector.

f64x4

Constants

FRAC_PI_2

π/2

MINUS_INFINITY_DB

MINUS_INFINITY_GAIN

PI

Archimedes’ constant (π)

TAU

The full circle constant (τ)

Functions

random_f64

A wrapper function around the random function that avoids the need for specifying a type in the case that it cannot be inferred. The primary purpose for this is to simplify the random API for new rust users.

random_range

A function for generating a random value within the given range.

Type Aliases

NoteHandlerRef