Module creative_coding_project::util::general

General-purpose utility functions.

Functions

db_to_level

Calculates the linear power level from amplitude as decibels.

dot

epsilon_eq

Returns true if value is equal to target, with a tolerance of ±f64::EPSILON.

freq_lin_from_log

The inverse of freq_log_norm().

freq_log_norm

Returns a normalised value representing the logarithmic value of a frequency based on the current sample rate.

freq_to_note

Calculates the MIDI note value of the provided frequency value.

interleave_dvec2_to_f64

Returns a vector of interleaved elements from the input, i.e. 0.x, 0.y, 1.x, 1.y, ...

interleave_vec2_to_f32

Returns a vector of interleaved elements from the input, i.e. 0.x, 0.y, 1.x, 1.y, ...

lanczos_kernel

Returns a vector containing points of a Lanczos kernel. a_factor is the “a” variable in the kernel calculation. Only holds enough points to represent each lobe. Returns 4 * a_factor + 1 elements (when trim_zeroes == false).

level_to_db

Calculates amplitude in decibels from a linear power level.

map

Maps a value from the provided input range to the provided output range.

map_f32

Maps a value from the provided input range to the provided output range.

normalize

Normalizes a value from a provided range.

normalize_f32

Normalizes a value from a provided range.

note_to_freq

Calculates the frequency value of the provided MIDI note value.

sample_length

Returns the length of one sample in seconds, based on the current sample rate.

scale

Scales a value to a provided range, assuming it is normalised.

scale_f32

Scales a value to a provided range, assuming it is normalised.

sinc

The unnormalised sinc function (sin(x) / x).

within_tolerance

Returns whether the absolute value of value is less than the provided tolerance value. Useful for checking approximate equality.