Function creative_coding_project::util::interp::cubic_hermite

pub fn cubic_hermite(
    p0: f64,
    p1: f64,
    p2: f64,
    p3: f64,
    t: f64,
    tension: f64,
    bias: f64
) -> f64

Interpolates between p1 and p2 based on the value of t using Hermite cubic interpolation, which requires four samples. This is more expensive than the Interp::cubic_catmull function, but allows for extra control via the tension and bias arguments.

t is clamped between 0 and 1.

tension: positive values increase curve “tension”, negative values reduce it. 0 is the default, “unaffected” value.

bias: positive values “skew” toward the last point, and negative values toward the first. 0 is the default, “unaffected” value.